Make read & write functions public for now.
This allows an external (replication) client to communicate with a MySQL
server over network by sending additional commands to the server. The
binlog replication stream is currently not supported by this crate, by
exposing these functions a dedicated client can re-use most of the
existing functionality and add their own event handling on top of it.